### 2019-10-21 #1817 @melissawm
@bocklund would you be willing to review this paper for JOSS? Thanks!
