### 2019-10-21 #1824 @whedon
[REVIEW]: pyscal : A python module for structural analysis of atomic environments
**Submitting author:** @srmnitc (<a href="http://orcid.org/0000-0002-6776-1213">Sarath Menon</a>)
**Repository:** <a href="https://github.com/srmnitc/pyscal" target ="_blank">https://github.com/srmnitc/pyscal</a>
**Version:** v2.0.1
