### 2019-10-24 srmnitc/pyscal#31 @bocklund
NumPy and matplotlib should added as requirements
I'm working on my JOSS review for this project (at https://github.com/openjournals/joss-reviews/issues/1824)

I installed `pyscal` into an Anaconda environment on macOS, but NumPy and matplotlib were not installed into that environment and I had to install them manually. I think this is caused because NumPy is only a `build` requirement, not a `run` requirement in `conda-recipe/meta.yaml`. When `pyscal` is installed via `conda`, neither NumPy nor matplotlib are installed.

This can be checked by running `conda search -c pyscal pyscal --info`, which only shows Python as a dependency.
