### 2019-10-24 srmnitc/pyscal#31 @srmnitc
Thanks for the feedback. I will fix this issue soon and get back to you.
