### 2019-10-25 srmnitc/pyscal#32 @bocklund
Code quality suggestions
Here are some suggestions that I have as I am working through as part of my review at https://github.com/openjournals/joss-reviews/issues/1824.

I don't consider any of these to be blocking for the review, so feel free to act on or not to act on my suggestions as you are able!

- [ ] Don't shadow names of built in functions (`Atom` having an `id` attribute) or standard library modules (the `pickle.py` module and all the tests importing the `sys` module, but then the API uses `sys = pc.System()`, overwriting that name) ([reference](http://python-notes.curiousefficiency.org/en/latest/python_concepts/import_traps.html#the-name-shadowing-trap))
- [ ] Don't use mutable parameters as default function arguments, examples: ["customkeys=[]"](https://github.com/srmnitc/pyscal/blob/9b46fe531caaac8adefec6218688022a09fc5f95/src/pyscal/traj_process.py#L13) and ["repetitions=[1, 1, 1]"](https://github.com/srmnitc/pyscal/blob/9b46fe531caaac8adefec6218688022a09fc5f95/src/pyscal/crystal_structures.py#L8) ([reference](https://docs.python-guide.org/writing/gotchas/#mutable-default-arguments))
