### 2019-10-25 srmnitc/pyscal#33 @bocklund
Documentation suggestions
The documentation looks great. I can tell you put a lot of hard work into it.

Here are some suggestions that I have as I am working through as part of my review at https://github.com/openjournals/joss-reviews/issues/1824.

I don't consider any of these to be blocking for the review, so feel free to act on or not to act on my suggestions as you are able!

* It's a little unclear in the documentation what features are supported in terms of `sys.read_inputfile`. For example, the standard input file I dumped from lammps (the simplest example in their documentation: `dump myDump all atom 100 dump.atom` [link](https://lammps.sandia.gov/doc/dump.html)) gives me fields of `ITEM: ATOMS id type xs ys zs` with scaled coordinates. This raises the following error with `sys.read_inputfile`. It seems like the scaled parameters are not supported, but this is neither documented nor is a good error message reported that might suggest that I've made a mistake. The same can be said for POSCAR files (Cartesian vs. Direct coordinates).

```python
KeyError                                  Traceback (most recent call last)
<ipython-input-6-4a42d161e44d> in <module>
----> 1 sys.read_inputfile('dump0.atom')

~/anaconda3/envs/joss_pyscal_env/lib/python3.6/site-packages/pyscal/core.py in read_inputfile(self, filename, format, frame, compressed, customkeys)
    166
    167             elif os.path.exists(filename):
--> 168                 atoms, boxdims, box, triclinic = ptp.read_lammps_dump(filename, compressed=compressed, check_triclinic=True, box_vectors=True, customkeys=customkeys)
    169                 self.atoms = atoms
    170                 self.box = boxdims

~/anaconda3/envs/joss_pyscal_env/lib/python3.6/site-packages/pyscal/traj_process.py in read_lammps_dump(infile, compressed, check_triclinic, box_vectors, customkeys)
    129             idd = int(raw[headerdict["id"]])
    130             typ = int(raw[headerdict["type"]])
--> 131             x = float(raw[headerdict["x"]])
    132             y = float(raw[headerdict["y"]])
    133             z = float(raw[headerdict["z"]])

KeyError: 'x'
```

* It seems like the examples on the documentation website are built up of tutorial notebooks. Since the tutorial use files from the examples directory, it could be useful to mention where the files come from (I didn't know at first how they should be found) and to provide a way for users to get and run the example files if they are not GitHub savvy.
