### 2019-10-25 #1824 @bocklund
I just finished my first pass on the review. I have a couple outstanding items to address:

**Installation:** Does installation proceed as outlined in the documentation? - I opened https://github.com/srmnitc/pyscal/issues/31. After this is closed, I can check this box.

**State of the field:** Do the authors describe how this software compares to other commonly-used packages - This is missing from the documentation and the JOSS paper.
