### 2019-10-25 #1824 @srmnitc
@bocklund thanks a lot for the quick review. I will work on the issues soon. Just a quick clarification regarding the **State of the field** part. I know ideally it should be on the paper, but do you think adding it to documentation will suffice?
