### 2019-10-25 073c055
Update meta file for conda recipe to include numpy and matplotlib as run dependencies
