+++ conda-recipe/meta.yaml
@@ -10,10 +10,11 @@ requirements:
   build:
     - python
     - setuptools
-    - numpy
 
   run:
     - python
+    - numpy
+    - matplotlib
 
 build:
     script_env:
