### 2019-10-26 srmnitc/pyscal#34 @lucydot
Links in documentation
Part of JOSS review process at openjournals/joss-reviews#1824 .
I think the documentation is great! Just a few minor points.

I worked through the notebooks on binder and spotted a couple of points where linking to further information might be useful.

- A link to Jupyter Notebook documentation would be useful for users who are not familiar with notebooks.

- In the solid/liquid example notebook a link to https://pyscal.readthedocs.io/en/latest/solidliquid.html would be useful.
