### 2019-10-26 srmnitc/pyscal#34 @lucydot
Another link and clarification that could be added to the basic example notebook would be that values should not be assigned to `sys.atoms[0]` , as highlighted elsewhere in the documentation ([link](https://pyscal.readthedocs.io/en/latest/example1basic.html#combining-system-and-atom)). The notebook example seems to suggest you can use it for getting values (`atom = sys.atoms[0]`), which is contradicted in the previous link.
