### 2019-10-26 #1824 @lucydot
Hi @srmnitc - I just finished my first look over. The documentation is really nice, and I like the use of binder for someone who wants to get a quick feel for the code.

**Installation:** I have the same issue as reported in srmnitc/pyscal#31 when installing through conda. I can confirm installation runs smoothly using `python setup.py install` in a conda environment as outlined in the documentation [here](https://pyscal.readthedocs.io/en/latest/gettingstarted.html#using-a-conda-environment).

**State of the field:** I don't want to call whether this is ok in the broader documentation, perhaps the editor @melissawm could clarify? It's listed under `Software paper` in the checklist, but I think it is a new addition to the checklist and there isn't explicit mention of it in the JOSS ReadTheDocs ([link](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)). It is stated that there should be `"A list of key references, including to other software addressing related needs"` - you reference Lechner 2016, though perhaps there is other software that could be referenced in addition to this. For me, it also links to the statement of need - what does your code enable that existing codes do not?

**Documentation:** I have a few suggestions regarding the documentation; these are just ideas that would not block paper acceptance. I've raised an issue in the repo: srmnitc/pyscal#34.
