### 2019-10-26 #1824 @srmnitc
@lucydot thanks for the review and the comments. I think I can add few sentences in the paper about other codes. There are few other python modules, but none of them seems to be maintained/documented, and lacks all of the features this module has. Bigger tools like LAMMPS, for example, can also do this, but the target is different. We are looking at mostly post-processing of data. But I will add few sentences in the paper regarding this. Once again, thanks for the quick review.
