### 2019-10-26 c81fb02
modify meta to include numpy and matplotlib as conda dependencies
