+++ conda-recipe/meta.yaml
@@ -8,13 +8,15 @@ source:
 
 requirements:
   build:
+    - cmake
     - python
     - setuptools
+    - numpy
 
   run:
     - python
     - numpy
-    - matplotlib
+    - matplotlib>=3
 
 build:
     script_env:
@@ -25,14 +27,17 @@ test:
     - pyscal
 
 about:
-  home: https://github.com/srmnitc/pyscal
-  license_file: ../LICENSE
+  home: https://pyscal.com
+  license: GPL-3.0
+  license_family: GPL
+  license_file: LICENSE
   summary: "Python library written in C++ for calculation of local atomic structural environment"
   description: |
      pyscal is a python module for the calculation of local atomic structural environments including Steinhardt's bond orientational order parameters
      during post-processing of atomistic simulation data. The core functionality of pyscal is written in C++ with python wrappers using pybind11 which
      allows for fast calculations with possibilities for easy expansion in python.
   doc_url: https://pyscal.com
+  dev_url: https://github.com/srmnitc/pyscal/ 
 
 extra:
   recipe-maintainers:
