### 2019-10-26 srmnitc/pyscal#31 @srmnitc
c81fb029e684b7a4cb5e964203cc42386c1f9216 should fix this issue. Both numpy and matplotlib has been added as dependencies.
