### 2019-10-26 #1824 @srmnitc
@lucydot @bocklund I have added both numpy and matplotlib as dependencies. srmnitc/pyscal#31 has been closed.
