+++ joss/paper.bib
@@ -128,3 +128,11 @@
     year      =   {2012},
     pages     =   {234107}
 }
+
+@misc{Leocmach2017,
+    title={Pyboo: A Python package to compute bond orientational order parameters},
+    author={Leocmach, A},
+    year={2017},
+    doi={10.5281/zenodo.1066568},
+    url={https://github.com/MathieuLeocmach/pyboo}
+}
