+++ joss/paper.md
@@ -36,7 +36,7 @@ Furthermore, Voronoi tessellation can be employed for a parameter-free definitio
 Higher exponents of the facet area to weight the bond orientational order parameters can further improve the recognition of common crystal phases [@Haeberle:2019].
 
 Originally inspired by the C++ code [``BondOrderAnalysis``](https://homepage.univie.ac.at/wolfgang.lechner/bondorderparameter.html) [@Lechner:2010] for the calculation of bond order parameters, ``pyscal`` is a python module designed for the computation of local structural order parameters  during post-processing of atomistic simulation data.
-In addition to the calculation of Steinhardt parameters which is offered by tools like [``BondOrderAnalysis``](https://homepage.univie.ac.at/wolfgang.lechner/bondorderparameter.html) and python module ``pyboo`` [@Leocmach2017], ``pyscal`` brings together the various approaches for structure identification discussed above in a single python module. While python offers the advantages of extensibility and flexibility, the module ensures the speed and efficiency of the calculations by using a core code written in C++, which is  ported to python using ``pybind11`` [@Jakob:2016]. The ``pyscal`` module includes the following functionalities:   
+In addition to the calculation of Steinhardt parameters which is offered by tools like [``BondOrderAnalysis``](https://homepage.univie.ac.at/wolfgang.lechner/bondorderparameter.html) and ``pyboo`` [@Leocmach2017], ``pyscal`` brings together the various approaches for structure identification discussed above in a single python module. While python offers the advantages of extensibility and flexibility, the module ensures the speed and efficiency of the calculations by using a core code written in C++, which is  ported to python using ``pybind11`` [@Jakob:2016]. The ``pyscal`` module includes the following functionalities:   
 
  * calculation of Steinhardt bond orientational order parameters and their averaged version.
  * weighted bond order parameters using face areas of Voronoi polyhedra; the calculation of the Voronoi polyhedra is  enabled using the Voro++ code [@Rycroft:2009] integrated into ``pyscal``.
