### 2019-10-26 #1824 @srmnitc
@bocklund  @lucydot  @melissawm,  About **State of the field**, 661b6770c62dc2f5e91b066a28dcbb63f785cfbe has an updated version of paper that references another [python module](https://pyboo.readthedocs.io/en/latest/) has some overlapping functionality. This is the only one I could find with documentation and a doi. [LAMMPS](https://lammps.sandia.gov/) and [PLUMED](https://www.plumed.org/) can also calculate the basic version of Steinhardt parameters, but the primary focus is completely different, so I think they will not fall in the same category as this module.
