+++ tests/test_q8.py
@@ -1,5 +1,5 @@
 import pytest
-import os,sys,inspect
+import os
 import numpy as np
 import pyscal.core as pc
 import pyscal.crystal_structures as pcs
