### 2019-10-26 50ebe76
minor changes to doc to reflect namechange of pickle module
