### 2019-10-26 srmnitc/pyscal#32 @srmnitc
Thanks for the great suggestion. I have fixed most of the above issues in the commit 50ebe76d9ceda5c4dbacdac118c2534c6c9f02ce except for ``Atom`` having an ``id`` attribute. I think I will keep this as it is for now so that the keyword makes sense to the users.
