### 2019-10-26 18531f6
minor bug fix in traj process
