### 2019-10-26 f40ef6c
improve reading loop in traj process to remove unnecessary comparison
