### 2019-10-26 ce97064
add scaled coordinates for lammps
