### 2019-10-26 9f53ff6
add scaled test file
