### 2019-10-26 srmnitc/pyscal#33 @srmnitc
Commit 9f53ff6cb8dc91b30de17ca78f7649f97d68459a adds scaled coordinate support to lammps dump file format. Additionally if either regular or scaled coordinate headers are not found, it throws an error message.
