### 2019-10-26 d577ccb
fix #33 and #34
