+++ docs/source/examples.rst
@@ -4,11 +4,16 @@ Examples
 Different examples illustrating the functionality of this module is provided in this section. Interactive versions of these
 examples can be launched `here <https://mybinder.org/v2/gh/srmnitc/pybop/master?filepath=examples%2F>`_.
 
+Examples are designed to run on `jupyter notebooks <https://jupyter.org>`_. A good tutorial on jupyter notebooks can
+be found `here <https://www.datacamp.com/community/tutorials/tutorial-jupyter-notebook>`_. All files used in the examples
+can be found in ``examples/`` folder of the main `repository <https://github.com/srmnitc/pyscal>`_. A copy of the repository
+can be downloaded from `here <https://pyscal.readthedocs.io/en/latest/download.html>`_.
+
 Basic examples
 --------------
 
 .. toctree::
-    
+
     example1basic
     example2basic
 
@@ -20,6 +25,3 @@ Advanced examples
 	example3basic
 	example_clustering
 	pickling
-
-
-    
