### 2019-10-26 srmnitc/pyscal#33 @srmnitc
Commit d577ccbaa6e3baa47836e7135989e6702ac8c580 add information in examples section about location of files used in the examples. Thanks for the suggestions, I am closing this issue now.
