### 2019-10-26 srmnitc/pyscal#34 @srmnitc
Thanks for the suggestions! I have implemented the recommendations in commit d577ccbaa6e3baa47836e7135989e6702ac8c580. I am closing this issue now.
