### 2019-10-27 a4bbce7
Set up CI with Azure Pipelines
[skip ci]
