+++ azure-pipelines.yml
@@ -1,0 +1,39 @@
+#add the different platforms to test
+#in case of windows the following line has to be added
+#windows: 'vs2017-win2016'
+
+variables:
+  linux: 'ubuntu-16.04'
+  mac: 'macos-10.13'
+
+#do not include branches as of now
+#trigger:
+#- master
+
+jobs:
+- job: 'Linux'
+  pool:
+    vmImage: $(linux)
+  
+  strategy:
+    matrix:
+      Python27:
+        python.version: '2.7'
+      Python37:
+        python.version: '3.7'
+    maxParallel: 4
+  
+  steps:
+    - if [ "$AGENT_OS" = "linux" ]; then OS=Linux-x86_64; else OS=MacOSX-x86_64; fi   
+    - wget -O miniconda.sh https://repo.continuum.io/miniconda/Miniconda${CONDA:0:1}-latest-$OS.sh
+    - bash miniconda.sh -b -p $HOME/miniconda
+    - export PATH="$HOME/miniconda/bin:$PATH"
+    - conda config --set always_yes yes --set changeps1 no
+    - conda config --add channels conda-forge
+    - conda install -c conda-forge libiconv
+    - conda install -c anaconda setuptools
+    - conda update --yes conda
+    - conda install -q conda-build
+    - conda create -q -n test-environment python=$CONDA pip
+    - source activate test-environment
+  
