### 2019-10-27 46b5bac
Update azure-pipelines.yml for Azure Pipelines
