+++ azure-pipelines.yml
@@ -24,10 +24,8 @@ jobs:
     maxParallel: 4
   
   steps:
-    - if [ "$AGENT_OS" = "linux" ]; then OS=Linux-x86_64; else OS=MacOSX-x86_64; fi   
-    - wget -O miniconda.sh https://repo.continuum.io/miniconda/Miniconda${CONDA:0:1}-latest-$OS.sh
-    - bash miniconda.sh -b -p $HOME/miniconda
-    - export PATH="$HOME/miniconda/bin:$PATH"
+    - bash: echo "##vso[task.prependpath]$CONDA/bin"
+      displayName: Add conda to PATH
     - conda config --set always_yes yes --set changeps1 no
     - conda config --add channels conda-forge
     - conda install -c conda-forge libiconv
