### 2019-10-27 e17c291
Update azure-pipelines.yml for Azure Pipelines
