### 2019-10-27 3e84d65
Update azure-pipelines.yml for Azure Pipelines
