### 2019-10-27 c0522aa
Update azure-pipelines.yml for Azure Pipelines
