### 2019-10-27 3b07b21
Update azure-pipelines.yml for Azure Pipelines
