### 2019-10-27 3af54e5
Update azure-pipelines.yml for Azure Pipelines
