### 2019-10-27 0ddfc2a
Update azure-pipelines.yml for Azure Pipelines
