+++ azure-pipelines.yml
@@ -34,5 +34,10 @@ jobs:
         conda install -c anaconda setuptools
         conda update --yes conda
         conda install -q conda-build
-        conda install --yes --quiet --name conenv scipy
-      displayName: Setup Anaconda Env 
+      displayName: Setup Anaconda Environment
+
+    - bash: |
+        conda activate conenv
+        conda install -c conda-forge --file requirements_dev.txt
+        pip install -e .
+      displayName: Install package   
