### 2019-10-27 7e7d18d
Update azure-pipelines.yml for Azure Pipelines
