### 2019-10-27 86cf67c
azure add test
