### 2019-10-27 5ce0ae1
azure add test
