### 2019-10-27 45cea8a
azure add test
