+++ azure-pipelines.yml
@@ -46,4 +46,11 @@ jobs:
         source activate conenv
         pytest --cov-report term-missing --cov=pyscal tests/
         codecov
-      displayName: Run tests
+     displayName: Run tests
+
+    - bash: |
+        and(succeeded(), eq(variables['Build.SourceBranch'], 'refs/heads/master'))
+        conda install conda-build
+        conda install anaconda-client
+        bash conda-recipe/conda_upload.sh
+      displayName: Upload to conda
