### 2019-10-27 fe17bd0
azure add conda build 2
