+++ azure-pipelines.yml
@@ -50,7 +50,13 @@ jobs:
 
     - bash: |
         and(succeeded(), eq(variables['Build.SourceBranch'], 'refs/heads/master'))
+        source activate conenv
         conda install conda-build
         conda install anaconda-client
-        bash conda-recipe/conda_upload.sh
+        mkdir ~/conda-bld
+        conda config --set anaconda_upload no
+        export CONDA_BLD_PATH=~/conda-bld
+        export VERSION=`date +%Y.%m.%d`
+        conda build .
+        anaconda -t $CONDA_UPLOAD_TOKEN upload -u pyscal -l main $CONDA_BLD_PATH/*/pyscal-`date +%Y.%m.%d`-*.tar.bz2 --force
       displayName: Upload to conda
