### 2019-10-27 e937fbe
azure add conda build - condition
