+++ azure-pipelines.yml
@@ -49,7 +49,6 @@ jobs:
       displayName: Run tests
 
     - bash: |
-        and(succeeded(), eq(variables['Build.SourceBranch'], 'refs/heads/master'))
         source activate conenv
         conda install conda-build
         conda install anaconda-client
@@ -59,4 +58,5 @@ jobs:
         export VERSION=`date +%Y.%m.%d`
         conda build .
         anaconda -t $CONDA_UPLOAD_TOKEN upload -u pyscal -l main $CONDA_BLD_PATH/*/pyscal-`date +%Y.%m.%d`-*.tar.bz2 --force
+      condition: and(succeeded(), eq(variables['Build.SourceBranch'], 'refs/heads/master'))
       displayName: Upload to conda
