### 2019-10-27 97df4df
azure add conda build - condition
