### 2019-10-27 640053e
azure add conda build - condition
