### 2019-10-27 f60b5af
azure add conda build - condition
