### 2019-10-27 2a70809
azure add conda build - condition
