+++ azure-pipelines.yml
@@ -57,6 +57,6 @@ jobs:
         export CONDA_BLD_PATH=~/conda-bld
         export VERSION=`date +%Y.%m.%d`
         conda build .
-        anaconda -t $CONDA_UPLOAD_TOKEN upload -c srmnitc -l main $CONDA_BLD_PATH/*/pyscal-`date +%Y.%m.%d`-*.tar.bz2 --force
+        anaconda upload -t ${CONDA_UPLOAD_TOKEN} -l main ${CONDA_BLD_PATH}/*/pyscal-`date +%Y.%m.%d`-*.tar.bz2 --force
       condition: and(succeeded(), eq(variables['Build.SourceBranch'], 'refs/heads/azure-pipelines'))
       displayName: Upload to conda
