### 2019-10-27 ae9c614
Update azure-pipelines.yml for Azure Pipelines
