### 2019-10-27 d60cd3f
Update azure-pipelines.yml try macos
