### 2019-10-27 7f8e8b8
Update azure-pipelines.yml try macos
