+++ azure-pipelines.yml
@@ -61,7 +61,7 @@ jobs:
       condition: and(succeeded(), eq(variables['Build.SourceBranch'], 'refs/heads/azure-pipelines'))
       displayName: Upload to conda
 
-jobs:
+
 - job: 'MacOS'
   pool:
     vmImage: $(mac)
