### 2019-10-27 0020e4c
Update azure-pipelines.yml for Azure Pipelines
