+++ azure-pipelines.yml
@@ -77,6 +77,7 @@ jobs:
   steps:
     - bash: |
         echo "##vso[task.prependpath]$CONDA/bin"
+        sudo chown -R $USER $CONDA
         conda create --yes --quiet --name conenv
         source activate conenv
         conda config --set always_yes yes --set changeps1 no
