### 2019-10-27 d8ff51e
Update azure-pipelines.yml for Azure Pipelines
