### 2019-10-27 e0110e4
Update azure-pipelines.yml for Azure Pipelines
