### 2019-10-27 9c2b479
Update azure-pipelines.yml for Azure Pipelines
