### 2019-10-27 5bb1d96
Update azure-pipelines.yml for Azure Pipelines
