### 2019-10-27 9862cf9
finish azure pipeline config
