### 2019-10-27 744ad38
finalise azure pipeline configuration
