+++ requirements.txt
@@ -1,1 +1,1 @@
-matplotlib==3.*
+matplotlib
