### 2019-10-27 c9e4865
minor modification in docs
