+++ docs/source/index.rst
@@ -6,10 +6,10 @@
 pyscal- A python module for structural analysis of atomic environments
 ======================================================================
 
-.. image:: https://travis-ci.com/srmnitc/pyscal.svg?branch=master
-    :target: https://travis-ci.com/srmnitc/pyscal
+.. image:: https://dev.azure.com/sarathrmenon/pyscal/_apis/build/status/srmnitc.pyscal?branchName=master
+    :target: https://dev.azure.com/sarathrmenon/pyscal/_build/latest?definitionId=1&branchName=master
     :width: 13%
-
+    
 .. image:: https://codecov.io/gh/srmnitc/pyscal/branch/master/graph/badge.svg
   :target: https://codecov.io/gh/srmnitc/pyscal
   :width: 15 %
