### 2019-10-28 #1824 @lucydot
@srmnitc  - fab, I can confirm that tests pass after install via conda. I'm also happy to tick off state of the field. I'd like another day or two to play around with the code (rather than just run through the examples given) before signing off - Lucy
