### 2019-10-28 752ba48
two major changes in calculation of solid atoms - 1. A variable q value can be used 2. The use of a fraction for bonds value
