+++ src/pyscal/system.cpp
@@ -28,6 +28,8 @@ System::System(){
     triclinic = 0;
     alpha = 1;
     voronoiused = 0;
+    solidq = 6;
+    criteria = 0;
 
 }
 
@@ -1158,12 +1160,12 @@ double System::get_number_from_bond(int ti,int tj){
     realdotproduct = 0.0;
     imgdotproduct = 0.0;
 
-    for (int mi = 0;mi < 13;mi++){
+    for (int mi = 0;mi < 2*solidq+1 ;mi++){
 
-        sumSquareti += atoms[ti].realq[4][mi]*atoms[ti].realq[4][mi] + atoms[ti].imgq[4][mi] *atoms[ti].imgq[4][mi];
-        sumSquaretj += atoms[tj].realq[4][mi]*atoms[tj].realq[4][mi] + atoms[tj].imgq[4][mi] *atoms[tj].imgq[4][mi];
-        realdotproduct += atoms[ti].realq[4][mi]*atoms[tj].realq[4][mi];
-        imgdotproduct  += atoms[ti].imgq[4][mi] *atoms[tj].imgq[4][mi];
+        sumSquareti += atoms[ti].realq[solidq-2][mi]*atoms[ti].realq[solidq-2][mi] + atoms[ti].imgq[solidq-2][mi] *atoms[ti].imgq[solidq-2][mi];
+        sumSquaretj += atoms[tj].realq[solidq-2][mi]*atoms[tj].realq[solidq-2][mi] + atoms[tj].imgq[solidq-2][mi] *atoms[tj].imgq[solidq-2][mi];
+        realdotproduct += atoms[ti].realq[solidq-2][mi]*atoms[tj].realq[solidq-2][mi];
+        imgdotproduct  += atoms[ti].imgq[solidq-2][mi] *atoms[tj].imgq[solidq-2][mi];
     }
 
     connection = (realdotproduct+imgdotproduct)/(sqrt(sumSquaretj)*sqrt(sumSquareti));
@@ -1182,12 +1184,12 @@ double System::get_number_from_bond(Atom atom1,Atom atom2){
     realdotproduct = 0.0;
     imgdotproduct = 0.0;
 
-    for (int mi = 0;mi < 13;mi++){
+    for (int mi = 0;mi < 2*solidq+1 ;mi++){
 
-        sumSquareti += atom1.realQ6[mi]*atom1.realQ6[mi] + atom1.imgQ6[mi] *atom1.imgQ6[mi];
-        sumSquaretj += atom2.realQ6[mi]*atom2.realQ6[mi] + atom2.imgQ6[mi] *atom2.imgQ6[mi];
-        realdotproduct += atom1.realQ6[mi]*atom2.realQ6[mi];
-        imgdotproduct  += atom1.imgQ6[mi] *atom2.imgQ6[mi];
+        sumSquareti += atom1.realq[solidq-2][mi]*atom1.realq[solidq-2][mi] + atom1.imgq[solidq-2][mi] *atom1.imgq[solidq-2][mi];
+        sumSquaretj += atom2.realq[solidq-2][mi]*atom2.realq[solidq-2][mi] + atom2.imgq[solidq-2][mi] *atom2.imgq[solidq-2][mi];
+        realdotproduct += atom1.realq[solidq-2][mi]*atom2.realq[solidq-2][mi];
+        imgdotproduct  += atom1.imgq[solidq-2][mi] *atom2.imgq[solidq-2][mi];
     }
 
     connection = (realdotproduct+imgdotproduct)/(sqrt(sumSquaretj)*sqrt(sumSquareti));
@@ -1237,14 +1239,19 @@ int System::cluster_criteria(int ti, int criterium){
 
[…]
