### 2019-10-28 bad8219
add testing of angular params
