### 2019-10-29 2a34682
add calculation of chi params
