### 2019-10-29 73f5b1c
fix bug in chi params
