### 2019-10-29 607263e
fix bug in chi params
