### 2019-10-29 84c9dcb
fix bug in chi params
