+++ docs/source/conf.py
@@ -26,9 +26,9 @@ copyright = u'2019, Sarath Menon, Grisell Díaz Leines, Jutta Rogal'
 author = u'Sarath Menon, Grisell Díaz Leines, Jutta Rogal'
 
 # The short X.Y version
-version = u''
+version = u'2.0.1'
 # The full version, including alpha/beta/rc tags
-release = u'2.0.0'
+release = u'2.0.1'
 
 
 # -- General configuration ---------------------------------------------------
