### 2019-10-30 5dc15e5
add function to return cos values
