### 2019-10-30 f3837a8
add nlimit criteria for chi params
