### 2019-11-01 #1824 @melissawm
@srmnitc A few points on the citations:
- In the citation of *Auer, S. & Frenkel, D. (2005)*, the correct citation from (https://link.springer.com/chapter/10.1007%2Fb99429#citeas)[Springer] reads: "Cite this chapter as: Auer S., Frenkel D. Numerical Simulation of Crystal Nucleation in Colloids. In: Dr. Holm C., Prof. Dr. Kremer K. (eds) Advanced Computer Simulation. Advances in Polymer Science, vol 173. Springer, Berlin, Heidelberg"
- Just to make sure, when you cite pybind11 you cite it as *Jakob, W., Rhinelander, J., & Moldovan, D. (2016)*, but there are several other contributors listed on that project. Is that information something the authors recommend for citing?
- For the citation "Kresse, G., & Furthmüller, J. (1996a). Efficiency of ab-initio total energy calculations for
metals and semiconductors using a plane-wave basis set. Computational Materials Science, 6, 15.", please check the Volume, Issue and Pages number (I believe it should read Volume 6, Issue 1, Pages 15-50)
- In the citation to VORO++, I believe it should read "Chaos: An Interdisciplinary Journal of Nonlinear Science, 19, 041111"
- Please check the Volume and Number information for the citation to "Stukowski, A (2012)" (I believe from the website it should be Volume 20, Number 4)
