### 2019-11-01 #1824 @srmnitc
@melissawm I have not included the number for any of the citations. Should I include the number too?
