+++ joss/paper.bib
@@ -1,11 +1,12 @@
 
-@article{Auer:2005,
+@book{Auer:2005,
     author    =   {Auer, S. and Frenkel, D.},
-    title     =   {{Numerical simulation of crystal nucleation in colloids}},
-    journal   =   {Advances in Polymer Science},
+    editor    =   {Dr. Holm, Christian and Prof. Dr. Kremer, Kurt},
+    title     =   {Numerical {Simulation} of {Crystal} {Nucleation} in {Colloids}},
+    publisher =   {Springer Berlin Heidelberg},
+    booktitle =   {Advanced Computer Simulation. Advances in Polymer Science},
     doi       =   {10.1007/b99429},
     volume    =   {173},
-    pages     =   {149},
     year      =   {2005}
 }
 
@@ -96,6 +97,7 @@
     journal   =   {Chaos: An Interdisciplinary Journal of Nonlinear Science},
     doi       =   {10.1063/1.3215722},
     volume    =   {19},
+    pages     =   {041111},
     year      =   {2009}
 }
 
