+++ joss/paper.bib
@@ -4,7 +4,7 @@
     editor    =   {Dr. Holm, Christian and Prof. Dr. Kremer, Kurt},
     title     =   {Numerical {Simulation} of {Crystal} {Nucleation} in {Colloids}},
     publisher =   {Springer Berlin Heidelberg},
-    booktitle =   {Advanced Computer Simulation. Advances in Polymer Science},
+    subtitle =   {Advanced Computer Simulation. Advances in Polymer Science},
     doi       =   {10.1007/b99429},
     volume    =   {173},
     year      =   {2005}
