### 2019-11-01 #1824 @srmnitc
@melissawm  I have corrected most of the issues. I have not included number/issue in any of the citations. For pybind11, I have used the proposed way of citation by the developers. Also for page numbers, I have only used the first page number. Once again, thanks for the suggestions!
