### 2019-11-01 #1824 @melissawm
@openjournals/joss-eics I believe this is ready for final acceptance.

Thanks, @srmnitc !
