### 2019-11-01 #1824 @danielskatz
@srmnitc - can you remove the space before the colon in the paper.md and the zenodo repository title?  This should be "pyscal: A ..."
