### 2019-11-01 fa88ce2
Update paper.md
