### 2019-11-01 b615b23
a few more small changes
