+++ joss/paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 The structural characterization of local atomic environments is essential in our understanding and design of materials with customized properties.  Atomistic simulations have become a powerful tool to provide insight into the structural patterns and mechanisms that signify transformations between different crystalline and liquid phases, or the formation and dynamics of point and extended defects.  The development and availability of methods that provide a reliable and accurate structural analysis of atomistic simulation data constitute an indispensable task that continues to be an active field of research in materials science.
 
 Several techniques have been proposed over the years to identify local structural motifs in crystalline materials [@Stukowski:2012].
-To distinguish solid- and liquid-like environments Auer and Frenkel [@Auer:2005] introduced a criterion that measures the structural correlations between a particle and its neighbors using the bond orientational order parameters suggested by Steinhardt et al. [@Steinhardt:1983].
+To distinguish solid- and liquid-like environments, Auer and Frenkel [@Auer:2005] introduced a criterion that measures the structural correlations between a particle and its neighbors using the bond orientational order parameters suggested by Steinhardt et al. [@Steinhardt:1983].
 The Steinhardt parameters are based on the spherical harmonics in the local neighborhood of a particle within a cutoff distance, and are rotationally and translationally invariant. They have been used extensively for the identification of crystal structures [@Mickel:2013 and references within].
 At finite temperatures the Steinhardt parameters for a given crystal structure exhibit a certain distribution of values instead of a single one due to thermal vibrations of the atoms.  This can lead to an overlap in the distributions for different phases which makes an accurate identification difficult or even impossible.
 An averaged version of the Steinhardt bond order parameters introduced by Lechner and Dellago [@Lechner:2008] significantly reduces this overlap and improves the characterization of common crystal structures.
@@ -35,8 +35,8 @@ An additional complication at finite temperatures is the use of a fixed cutoff d
 Furthermore, Voronoi tessellation can be employed for a parameter-free definition of the local neighborhood of a particle. Mickel et al. showed [@Mickel:2013] that the identification of crystal structures can be improved by weighting the contribution of each neighbor to the Steinhardt parameters by the area of the Voronoi facet shared between the central atom and its corresponding neighbor.
 Higher exponents of the facet area to weight the bond orientational order parameters can further improve the recognition of common crystal phases [@Haeberle:2019].
 
-Originally inspired by the C++ code [``BondOrderAnalysis``](https://homepage.univie.ac.at/wolfgang.lechner/bondorderparameter.html) [@Lechner:2010] for the calculation of bond order parameters, ``pyscal`` is a python module designed for the computation of local structural order parameters  during post-processing of atomistic simulation data.
-In addition to the calculation of Steinhardt parameters which is offered by tools like [``BondOrderAnalysis``](https://homepage.univie.ac.at/wolfgang.lechner/bondorderparameter.html) and ``pyboo`` [@Leocmach2017], ``pyscal`` brings together the various approaches for structure identification discussed above in a single python module. While python offers the advantages of extensibility and flexibility, the module ensures the speed and efficiency of the calculations by using a core code written in C++, which is  ported to python using ``pybind11`` [@Jakob:2016]. The ``pyscal`` module includes the following functionalities:   
+Originally inspired by the C++ code [``BondOrderAnalysis``](https://homepage.univie.ac.at/wolfgang.lechner/bondorderparameter.html) [@Lechner:2010] for the calculation of bond order parameters, ``pyscal`` is a Python module designed for the computation of local structural order parameters  during post-processing of atomistic simulation data.
[…]
