### 2019-11-01 #1824 @danielskatz
Can you also merge https://github.com/srmnitc/pyscal/pull/38 ?
