### 2019-11-01 #1824 @srmnitc
@danielskatz merged
