### 2019-11-01 #1824 @danielskatz
sorry, I didn't check the bib entries at the same time - can you also merge https://github.com/srmnitc/pyscal/pull/39 ?
