### 2019-11-01 #1824 @danielskatz
Thanks to @lucydot & @bocklund for reviewing and @melissawm for editing!
And to @srmnitc for putting up with up with a bunch of little last minute changes!
