### 2020-07-09 #1824 @srmnitc
@arfon Hi Arfon, I am the author of the package in this review, which is now published. I had a quick question. We would like to change the license of the package from GNU GPL to BSD. It still remains open source, but it is for integration into a larger project. Is this okay from the perspective of JOSS? Sorry about the trouble!
